Store: Support disable cache index header file.

+++ pkg/block/indexheader/binary_reader.go
@@ -5,6 +5,7 @@ package indexheader
 
 import (
 	"bufio"
+	"bytes"
 	"context"
 	"encoding/binary"
 	"hash"
@@ -69,57 +70,64 @@ type BinaryTOC struct {
 	PostingsOffsetTable uint64
 }
 
-// WriteBinary build index-header file from the pieces of index in object storage.
-func WriteBinary(ctx context.Context, bkt objstore.BucketReader, id ulid.ULID, filename string) (err error) {
+// WriteBinary build index header from the pieces of index in object storage, and cached in file if necessary.
+func WriteBinary(ctx context.Context, bkt objstore.BucketReader, id ulid.ULID, filename string) ([]byte, error) {
 	ir, indexVersion, err := newChunkedIndexReader(ctx, bkt, id)
 	if err != nil {
-		return errors.Wrap(err, "new index reader")
+		return nil, errors.Wrap(err, "new index reader")
 	}
-	tmpFilename := filename + ".tmp"
+	tmpFilename := ""
+	if filename != "" {
+		tmpFilename = filename + ".tmp"
+	}
 
 	// Buffer for copying and encbuffers.
 	// This also will control the size of file writer buffer.
 	buf := make([]byte, 32*1024)
-	bw, err := newBinaryWriter(tmpFilename, buf)
+	bw, err := newBinaryWriter(id, tmpFilename, buf)
 	if err != nil {
-		return errors.Wrap(err, "new binary index header writer")
+		return nil, errors.Wrap(err, "new binary index header writer")
 	}
 	defer runutil.CloseWithErrCapture(&err, bw, "close binary writer for %s", tmpFilename)
 
 	if err := bw.AddIndexMeta(indexVersion, ir.toc.PostingsTable); err != nil {
-		return errors.Wrap(err, "add index meta")
+		return nil, errors.Wrap(err, "add index meta")
 	}
 
 	if err := ir.CopySymbols(bw.SymbolsWriter(), buf); err != nil {
-		return err
+		return nil, err
 	}
 
-	if err := bw.f.Flush(); err != nil {
-		return errors.Wrap(err, "flush")
+	if err := bw.writer.Flush(); err != nil {
+		return nil, errors.Wrap(err, "flush")
 	}
 
 	if err := ir.CopyPostingsOffsets(bw.PostingOffsetsWriter(), buf); err != nil {
-		return err
+		return nil, err
 	}
 
-	if err := bw.f.Flush(); err != nil {
-		return errors.Wrap(err, "flush")
+	if err := bw.writer.Flush(); err != nil {
+		return nil, errors.Wrap(err, "flush")
 	}
 
 	if err := bw.WriteTOC(); err != nil {
-		return errors.Wrap(err, "write index header TOC")
+		return nil, errors.Wrap(err, "write index header TOC")
 	}
 
-	if err := bw.f.Flush(); err != nil {
-		return errors.Wrap(err, "flush")
+	if err := bw.writer.Flush(); err != nil {
+		return nil, errors.Wrap(err, "flush")
 	}
 
-	if err := bw.f.f.Sync(); err != nil {
-		return errors.Wrap(err, "sync")
+	if err := bw.writer.Sync(); err != nil {
+		return nil, errors.Wrap(err, "sync")
 	}
 
-	// Create index-header in atomic way, to avoid partial writes (e.g during restart or crash of store GW).
-	return os.Rename(tmpFilename, filename)
+	if tmpFilename != "" {
+		// Create index-header in atomic way, to avoid partial writes (e.g during restart or crash of store GW).
+		return nil, os.Rename(tmpFilename, filename)
+	}
+
+	return bw.Buffer(), nil
 }
 
 type chunkedIndexReader struct {
@@ -231,7 +239,7 @@ func (r *chunkedIndexReader) CopyPostingsOffsets(w io.Writer, buf []byte) (err e
 
 // TODO(bwplotka): Add padding for efficient read.
 type binaryWriter struct {
-	f *FileWriter
+	writer PosWriter
 
 	toc BinaryTOC
 
@@ -241,38 +249,47 @@ type binaryWriter struct {
 	crc32 hash.Hash
 }
 
-func newBinaryWriter(fn string, buf []byte) (w *binaryWriter, err error) {
-	dir := filepath.Dir(fn)
+func newBinaryWriter(id ulid.ULID, fn string, buf []byte) (w *binaryWriter, err error) {
+	var writer PosWriter
+	if fn != "" {
+		dir := filepath.Dir(fn)
 
-	df, err := fileutil.OpenDir(dir)
-	if os.IsNotExist(err) {
-		if err := os.MkdirAll(dir, os.ModePerm); err != nil {
+		df, err := fileutil.OpenDir(dir)
+		if os.IsNotExist(err) {
+			if err := os.MkdirAll(dir, os.ModePerm); err != nil {
+				return nil, err
+			}
+			df, err = fileutil.OpenDir(dir)
+		}
+		if err != nil {
+
 			return nil, err
 		}
-		df, err = fileutil.OpenDir(dir)
-	}
-	if err != nil {
 
-		return nil, err
-	}
+		defer runutil.CloseWithErrCapture(&err, df, "dir close")
 
-	defer runutil.CloseWithErrCapture(&err, df, "dir close")
+		if err := os.RemoveAll(fn); err != nil {
+			return nil, errors.Wrap(err, "remove any existing index at path")
+		}
 
-	if err := os.RemoveAll(fn); err != nil {
-		return nil, errors.Wrap(err, "remove any existing index at path")
-	}
-
-	// We use file writer for buffers not larger than reused one.
-	f, err := NewFileWriter(fn, len(buf))
-	if err != nil {
-		return nil, err
-	}
-	if err := df.Sync(); err != nil {
-		return nil, errors.Wrap(err, "sync dir")
+		// We use file writer for buffers not larger than reused one.
+		writer, err = NewFileWriter(fn, len(buf))
+		if err != nil {
+			return nil, err
+		}
+		if err := df.Sync(); err != nil {
+			return nil, errors.Wrap(err, "sync dir")
+		}
+	} else {
+		var err error
+		writer, err = NewMemoryWriter(id, len(buf))
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	w = &binaryWriter{
-		f: f,
+		writer: writer,
 
 		// Reusable memory.
 		buf:   encoding.Encbuf{B: buf},
@@ -283,7 +300,70 @@ func newBinaryWriter(fn string, buf []byte) (w *binaryWriter, err error) {
 	w.buf.PutBE32(MagicIndex)
 	w.buf.PutByte(BinaryFormatV1)
 
-	return w, w.f.Write(w.buf.Get())
+	return w, w.writer.Write(w.buf.Get())
+}
+
+type PosWriter interface {
+	Pos() uint64
+	Write(bufs ...[]byte) error
+	Buffer() []byte
+	Flush() error
+	Sync() error
+	Close() error
+}
+
+type MemoryWriter struct {
+	id  ulid.ULID
+	buf bytes.Buffer
+	pos uint64
+}
+
+// TODO(bwplotka): Added size to method, upstream this.
+func NewMemoryWriter(id ulid.ULID, size int) (PosWriter, error) {
+	var buf bytes.Buffer
+	return &MemoryWriter{
+		id:  id,
+		buf: buf,
+		pos: 0,
+	}, nil
+}
+
+func (mw *MemoryWriter) Pos() uint64 {
+	return mw.pos
+}
+
+func (mw *MemoryWriter) Write(bufs ...[]byte) error {
+	for _, b := range bufs {
+		n, err := mw.buf.Write(b)
+		mw.pos += uint64(n)
+		if err != nil {
+			return err
+		}
+		// For now the index file must not grow beyond 64GiB. Some of the fixed-sized
+		// offset references in v1 are only 4 bytes large.
+		// Once we move to compressed/varint representations in those areas, this limitation
+		// can be lifted.
+		if mw.pos > 16*math.MaxUint32 {
+			return errors.Errorf("%q exceeding max size of 64GiB", mw.id)
+		}
+	}
+	return nil
+}
+
+func (mw *MemoryWriter) Buffer() []byte {
+	return mw.buf.Bytes()
+}
+
+func (mw *MemoryWriter) Flush() error {
+	return nil
+}
+
+func (mw *MemoryWriter) Sync() error {
+	return nil
+}
+
+func (mw *MemoryWriter) Close() error {
+	return mw.Flush()
 }
 
 type FileWriter struct {
@@ -329,6 +409,10 @@ func (fw *FileWriter) Write(bufs ...[]byte) error {
 	return nil
 }
 
+func (fw *FileWriter) Buffer() []byte {
+	return nil
+}
+
 func (fw *FileWriter) Flush() error {
 	return fw.fbuf.Flush()
 }
@@ -365,6 +449,10 @@ func (fw *FileWriter) Close() error {
 	return fw.f.Close()
 }
 
+func (fw *FileWriter) Sync() error {
+	return fw.f.Sync()
+}
+
 func (fw *FileWriter) Remove() error {
 	return os.Remove(fw.name)
 }
@@ -373,16 +461,16 @@ func (w *binaryWriter) AddIndexMeta(indexVersion int, indexPostingOffsetTable ui
 	w.buf.Reset()
 	w.buf.PutByte(byte(indexVersion))
 	w.buf.PutBE64(indexPostingOffsetTable)
-	return w.f.Write(w.buf.Get())
+	return w.writer.Write(w.buf.Get())
 }
 
 func (w *binaryWriter) SymbolsWriter() io.Writer {
-	w.toc.Symbols = w.f.Pos()
+	w.toc.Symbols = w.writer.Pos()
 	return w
 }
 
 func (w *binaryWriter) PostingOffsetsWriter() io.Writer {
-	w.toc.PostingsOffsetTable = w.f.Pos()
+	w.toc.PostingsOffsetTable = w.writer.Pos()
 	return w
 }
 
@@ -394,17 +482,21 @@ func (w *binaryWriter) WriteTOC() error {
 
 	w.buf.PutHash(w.crc32)
 
-	return w.f.Write(w.buf.Get())
+	return w.writer.Write(w.buf.Get())
 }
 
 func (w *binaryWriter) Write(p []byte) (int, error) {
-	n := w.f.Pos()
-	err := w.f.Write(p)
-	return int(w.f.Pos() - n), err
+	n := w.writer.Pos()
+	err := w.writer.Write(p)
+	return int(w.writer.Pos() - n), err
+}
+
+func (w *binaryWriter) Buffer() []byte {
+	return w.writer.Buffer()
 }
 
 func (w *binaryWriter) Close() error {
-	return w.f.Close()
+	return w.writer.Close()
 }
 
 type postingValueOffsets struct {
@@ -458,21 +550,45 @@ type BinaryReader struct {
 
 // NewBinaryReader loads or builds new index-header if not present on disk.
 func NewBinaryReader(ctx context.Context, logger log.Logger, bkt objstore.BucketReader, dir string, id ulid.ULID, postingOffsetsInMemSampling int) (*BinaryReader, error) {
-	binfn := filepath.Join(dir, id.String(), block.IndexHeaderFilename)
-	br, err := newFileBinaryReader(binfn, postingOffsetsInMemSampling)
-	if err == nil {
-		return br, nil
+	if dir != "" {
+		binfn := filepath.Join(dir, id.String(), block.IndexHeaderFilename)
+		br, err := newFileBinaryReader(binfn, postingOffsetsInMemSampling)
+		if err == nil {
+			return br, nil
+		}
+
+		level.Debug(logger).Log("msg", "failed to read index-header from disk; recreating", "path", binfn, "err", err)
+
+		start := time.Now()
+		if _, err := WriteBinary(ctx, bkt, id, binfn); err != nil {
+			return nil, errors.Wrap(err, "write index header")
+		}
+
+		level.Debug(logger).Log("msg", "built index-header file", "path", binfn, "elapsed", time.Since(start))
+		return newFileBinaryReader(binfn, postingOffsetsInMemSampling)
+	} else {
+		buf, err := WriteBinary(ctx, bkt, id, "")
+		if err != nil {
+			return nil, errors.Wrap(err, "generate index header")
+		}
+
+		return newMemoryBinaryReader(buf, postingOffsetsInMemSampling)
+	}
+}
+
+func newMemoryBinaryReader(buf []byte, postingOffsetsInMemSampling int) (bw *BinaryReader, err error) {
+	r := &BinaryReader{
+		b:                           realByteSlice(buf),
+		c:                           nil,
+		postings:                    map[string]*postingValueOffsets{},
+		postingOffsetsInMemSampling: postingOffsetsInMemSampling,
 	}
 
-	level.Debug(logger).Log("msg", "failed to read index-header from disk; recreating", "path", binfn, "err", err)
-
-	start := time.Now()
-	if err := WriteBinary(ctx, bkt, id, binfn); err != nil {
-		return nil, errors.Wrap(err, "write index header")
+	if err := r.init(); err != nil {
+		return nil, err
 	}
 
-	level.Debug(logger).Log("msg", "built index-header file", "path", binfn, "elapsed", time.Since(start))
-	return newFileBinaryReader(binfn, postingOffsetsInMemSampling)
+	return r, nil
 }
 
 func newFileBinaryReader(path string, postingOffsetsInMemSampling int) (bw *BinaryReader, err error) {
@@ -493,12 +609,44 @@ func newFileBinaryReader(path string, postingOffsetsInMemSampling int) (bw *Bina
 		postingOffsetsInMemSampling: postingOffsetsInMemSampling,
 	}
 
+	if err := r.init(); err != nil {
+		return nil, err
+	}
+
+	return r, nil
+}
+
+// newBinaryTOCFromByteSlice return parsed TOC from given index header byte slice.
+func newBinaryTOCFromByteSlice(bs index.ByteSlice) (*BinaryTOC, error) {
+	if bs.Len() < binaryTOCLen {
+		return nil, encoding.ErrInvalidSize
+	}
+	b := bs.Range(bs.Len()-binaryTOCLen, bs.Len())
+
+	expCRC := binary.BigEndian.Uint32(b[len(b)-4:])
+	d := encoding.Decbuf{B: b[:len(b)-4]}
+
+	if d.Crc32(castagnoliTable) != expCRC {
+		return nil, errors.Wrap(encoding.ErrInvalidChecksum, "read index header TOC")
+	}
+
+	if err := d.Err(); err != nil {
+		return nil, err
+	}
+
+	return &BinaryTOC{
+		Symbols:             d.Be64(),
+		PostingsOffsetTable: d.Be64(),
+	}, nil
+}
+
+func (r *BinaryReader) init() (err error) {
 	// Verify header.
 	if r.b.Len() < headerLen {
-		return nil, errors.Wrap(encoding.ErrInvalidSize, "index header's header")
+		return errors.Wrap(encoding.ErrInvalidSize, "index header's header")
 	}
 	if m := binary.BigEndian.Uint32(r.b.Range(0, 4)); m != MagicIndex {
-		return nil, errors.Errorf("invalid magic number %x", m)
+		return errors.Errorf("invalid magic number %x", m)
 	}
 	r.version = int(r.b.Range(4, 5)[0])
 	r.indexVersion = int(r.b.Range(5, 6)[0])
@@ -506,18 +654,18 @@ func newFileBinaryReader(path string, postingOffsetsInMemSampling int) (bw *Bina
 	r.indexLastPostingEnd = int64(binary.BigEndian.Uint64(r.b.Range(6, headerLen)))
 
 	if r.version != BinaryFormatV1 {
-		return nil, errors.Errorf("unknown index header file version %d", r.version)
+		return errors.Errorf("unknown index header file version %d", r.version)
 	}
 
 	r.toc, err = newBinaryTOCFromByteSlice(r.b)
 	if err != nil {
-		return nil, errors.Wrap(err, "read index header TOC")
+		return errors.Wrap(err, "read index header TOC")
 	}
 
 	// TODO(bwplotka): Consider contributing to Prometheus to allow specifying custom number for symbolsFactor.
 	r.symbols, err = index.NewSymbols(r.b, r.indexVersion, int(r.toc.Symbols))
 	if err != nil {
-		return nil, errors.Wrap(err, "read symbols")
+		return errors.Wrap(err, "read symbols")
 	}
 
 	var lastName, lastValue []byte
@@ -543,7 +691,7 @@ func newFileBinaryReader(path string, postingOffsetsInMemSampling int) (bw *Bina
 			prevRng = index.Range{Start: int64(postingsOffset + postingLengthFieldSize)}
 			return nil
 		}); err != nil {
-			return nil, errors.Wrap(err, "read postings table")
+			return errors.Wrap(err, "read postings table")
 		}
 		if string(lastName) != "" {
 			prevRng.End = r.indexLastPostingEnd - crc32.Size
@@ -583,7 +731,7 @@ func newFileBinaryReader(path string, postingOffsetsInMemSampling int) (bw *Bina
 
 			return nil
 		}); err != nil {
-			return nil, errors.Wrap(err, "read postings table")
+			return errors.Wrap(err, "read postings table")
 		}
 		if lastName != nil {
 			if (valueCount-1)%postingOffsetsInMemSampling != 0 {
@@ -609,38 +757,14 @@ func newFileBinaryReader(path string, postingOffsetsInMemSampling int) (bw *Bina
 		}
 		off, err := r.symbols.ReverseLookup(k)
 		if err != nil {
-			return nil, errors.Wrap(err, "reverse symbol lookup")
+			return errors.Wrap(err, "reverse symbol lookup")
 		}
 		r.nameSymbols[off] = k
 	}
 
 	r.dec = &index.Decoder{LookupSymbol: r.LookupSymbol}
 
-	return r, nil
-}
-
-// newBinaryTOCFromByteSlice return parsed TOC from given index header byte slice.
-func newBinaryTOCFromByteSlice(bs index.ByteSlice) (*BinaryTOC, error) {
-	if bs.Len() < binaryTOCLen {
-		return nil, encoding.ErrInvalidSize
-	}
-	b := bs.Range(bs.Len()-binaryTOCLen, bs.Len())
-
-	expCRC := binary.BigEndian.Uint32(b[len(b)-4:])
-	d := encoding.Decbuf{B: b[:len(b)-4]}
-
-	if d.Crc32(castagnoliTable) != expCRC {
-		return nil, errors.Wrap(encoding.ErrInvalidChecksum, "read index header TOC")
-	}
-
-	if err := d.Err(); err != nil {
-		return nil, err
-	}
-
-	return &BinaryTOC{
-		Symbols:             d.Be64(),
-		PostingsOffsetTable: d.Be64(),
-	}, nil
+	return nil
 }
 
 func (r *BinaryReader) IndexVersion() (int, error) {
